Split Stripe service to interface and implementation
